requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2
django-graphql-jwt==0.3.4
django-heroku==0.3.1
flake8==4.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
mccabe==0.6.1
promise==2.3
psycopg==3.0.7
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
